Adds a secondary, stricter hill climb at the end
For hard scoring, for each token that isn't an english word, we add its length to the score.  This means that if two letters with very similar digram frequencies are swapped, this hill climb can fix that.

With this change, we can
 - Abort early when the score gets low enough, because the hard hill climb is enough to get us the rest of the way there.
 - In about 30 test runs, I got perfect decryption.

We don't use this hard version of the score for the normal key search, because it's very expensive to compute for each swap, but putting one round of it in at the end makes all the difference in the world.